requirements.txt
discord.py
pyfiglet
